Audit next meeting date script (#130)
This PR makes two changes to the next meeting date scraping logic:

1. If the scraper encounters an error, the error catcher rethrows the
error so that further Github Action steps are aborted and no commits are
made against the code repository.
2. If the next meeting date is not a valid `Date` object, an error is
thrown so that further Github Action steps are aborted and no commits
are made against the code repository.

It is not practical to catch malformed JSON objects on the application
end since Vite will throw an error on import that the application cannot
catch. This isn't an issue we have to worry about since we create the
object ourselves and write it using `JSON.stringify`, which is probably
more robust than anything we can come up with.