Parse applicants (#128)
This PR introduces scripts to extract applicant data from PDFs and
maintain an up-to-date JSON database of license applicants.

The main script, extract_entity.py, handles extracting detailed
applicant information from PDF files. It is designed to be integrated
into a GitHub Action that will run on a weekly schedule. This action
will process new PDFs (such as voting minutes) as they become available
to keep the data current.

To preload historical data, a one-time seeding script called
load_data.py is included and located in scripts/archive. This script
processes all PDFs placed in the scripts folder and stores the extracted
data in a JSON file.

I’ve written the scripts in Python, I hope that’s okay. I’d also
appreciate help verifying whether the extracted data is accurate. The
data.json file included in this PR contains data from September 26 to
June 6, so it's not fully up to date yet. The data.json file is sorted
first by date and then by entity number, which corresponds to the bullet
number assigned to each applicant in the voting minutes.

This PR is mainly to start a conversation on whether this approach makes
sense, or if there's a better way we should consider. I'm happy to
revise things based on your feedback!

---------

Co-authored-by: Will Worrell <129802295+willrequaworrell@users.noreply.github.com>
Co-authored-by: Matthew Clarke <4278212+MattClarke131@users.noreply.github.com>
Co-authored-by: Haile Xiao <haile.xiao@gmail.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Github Actions <hello@codeforboston.org>
Co-authored-by: PranshuBU <spranshu@bu.edu>